Remove broken (un-sourced) random images
The post at
https://web.wsu.edu/2015/06/25/intentional-acts-of-randomness/
intended to show off the random classes assigned by the WSU Spine.

The images used on that did not have any attribution, so I've removed
them from the CSS. In the future, if we keep any kind of random
class, we should write it up as a guide, not a post, so that we
can maintain its current-ness.